+++ packages/sampler/stories/moonstone-stories/CheckboxItem.js
@@ -3,7 +3,7 @@ import ToggleItem from '@enact/moonstone/ToggleItem';
 import Item, {ItemBase} from '@enact/moonstone/Item';
 import React from 'react';
 import {storiesOf, action} from '@kadira/storybook';
-import {withKnobs, boolean, text} from '@kadira/storybook-addon-knobs';
+import {withKnobs, boolean, select, text} from '@kadira/storybook-addon-knobs';
 
 import {mergeComponentMetadata} from '../../src/utils/propTables';
 
@@ -17,6 +17,7 @@ storiesOf('CheckboxItem')
 		() => (
 			<CheckboxItem
 				disabled={boolean('disabled', false)}
+				iconPosition={select('iconPosition', ['before', 'after'], 'before')}
 				inline={boolean('inline', false)}
 				onToggle={action('onToggle')}
 			>