+++ packages/sampler/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact Sampler, newest changes on the top.
 
+## [unreleased]
+
+### Added
+
+- `moonstone/CheckboxItem` knob to change the checkbox position (left vs right aligned)
+
 ## [1.0.0-beta.4] - 2017-03-10
 
 No significant changes.